Introduce `lib/ember_cli.rb`
So that the files and tests follow Ruby's naming idioms, move the
contents of `lib/ember-cli-rails.rb` to `lib/ember_cli.rb`, to match the
constant that the file contains.
